Renames rogue zeros function to new_data
