+++ src/Fields/Fields.jl
@@ -11,6 +11,11 @@ export
 using Oceananigans.Grids
 using Oceananigans.BoundaryConditions
 
+Base.zeros(FT, ::CPU, Nx, Ny, Nz) = zeros(FT, Nx, Ny, Nz)
+Base.zeros(FT, ::GPU, Nx, Ny, Nz) = zeros(FT, Nx, Ny, Nz) |> CuArray
+Base.zeros(arch, grid, Nx, Ny, Nz) = zeros(eltype(grid), arch, Nx, Ny, Nz)
+
+include("new_data.jl")
 include("field.jl")
 include("set!.jl")
 include("field_tuples.jl")
